完成拣货顺序

+++ src/views/carriage/cargosequence/index.vue
@@ -102,7 +102,8 @@
                                   </el-col> -->
                               </el-tab-pane>
                               <el-tab-pane label="操作日志" name="operational">
-                                    <system-log :modular="'PICK_ORDER'" :id="storeArmy.id"></system-log>
+                                    <!-- <system-log :modular="'PICK_ORDER'" :id="storeArmy.id"></system-log> -->
+                                    <system-log :modular="'PICK_ORDER'" :id="editProjectsInfo.schemeList.id"></system-log>
                               </el-tab-pane>
                           </el-tabs>
                       </template>
@@ -115,16 +116,12 @@
                     <el-table
                         lable="团点"
                         :data="storeOption"
+                        :default-sort = "{prop: 'idx', order: 'descending'}"
                         style="width: 100%;text-align:center"
                     >
-                     <!-- <el-table-column label="序号">
+                     <el-table-column label="序号" sortable prop="idx">
                           <template slot-scope="scope">
                               {{ scope.row.idx }}
-                          </template>
-                      </el-table-column> -->
-                      <el-table-column label="顺序">
-                          <template slot-scope="scope">
-                              {{ scope.row.storeOrder}}
                           </template>
                       </el-table-column>
                       <el-table-column label="团点">
@@ -132,16 +129,16 @@
                               <span style="color:#409EFF">{{ "[" + scope.row.storeCode + "]" + scope.row.storeName}}</span>
                           </template>
                       </el-table-column>
-                      <el-table-column label="配送中心">
+                      <el-table-column label="顺序">
                           <template slot-scope="scope">
-                              {{ '[' + scope.row.dcCode + ']' + scope.row.dcName }}
+                              {{ scope.row.storeOrder}}
                           </template>
                       </el-table-column>
-                      <el-table-column label="状态">
+                      <!-- <el-table-column label="状态">
                           <template slot-scope="scope">
                               {{ scope.row.status | statusType}}
                           </template>
-                      </el-table-column>
+                      </el-table-column> -->
                       <el-table-column label="操作" width="200">
                           <template slot-scope="scope">
                             <el-button
@@ -376,6 +373,7 @@ import systemLog from "@/components/systemLog.vue"
 export default {
   data() {
       return {
+        adjustOrderFlag: {},
         afterNum: 1,
         editProjectsObj: {
           code: '',
@@ -552,7 +550,6 @@ export default {
    },
    // 获取门店组下面所有的所有的门店
    getAllStoreByGrpId(id) {
-     console.log(id);
       CargosequenceService.getAllStoreByGrpId(id)
       .then((res) => {
           const storeOptionArr = res;
@@ -723,11 +720,12 @@ export default {
       // "updatorId": "string",
       // "updatorName": "string",
       // "version": "string"
+      this.adjustOrderFlag = obj;
       this.storeMesAll = {
-        adjustOrder: obj.idx,
+        adjustOrder: obj.storeOrder,
         grpId: obj.grpId,
         id: obj.id,
-        pickOrderId: obj.pickOrderId,
+        pickorderId: obj.pickorderId,
         remark: obj.remark,
         status: obj.status,
         storeCode: obj.storeCode,
@@ -737,22 +735,39 @@ export default {
         getByIdMesName: '[' + this.editProjectsInfo.schemeList.code + ']' + this.editProjectsInfo.schemeList.name,
         getGrpByIdMesName: '[' + this.storeArmy.code + ']' + this.storeArmy.name,
         mylength: this.storeOption.length,
-        curIdx: obj.idx
+        curIdx: obj.storeOrder
       }
    },
    Cancellation() {
     this.dialogFormVisible = false;
-    this.postAdjustOrder(this.storeMesAll);
-    this.getAllStoreByGrpId(this.storeMesAll.id);
-   },
+    this.adjustOrderFlag.adjustOrder = this.storeMesAll.adjustOrder;
+    CargosequenceService.adjustOrder(this.adjustOrderFlag)
+    .then((res) => {
+     this.schemeOrStore(this.changeActive, this.storeArmy, this.editProjectsInfo)
+      this.$message.success("调序成功");
+    }).catch((err) => {
+      this.$message.error("调序失败" + err.message)
+    });
+  },
    deleteItem(obj) {
-      CargosequenceService.deleteItem(obj.id, obj.version)
-      .then((res) => {
-        this.$message.success("删除成功");
-        this.getAllStoreByGrpId(obj.id);
-      }).catch((err) => {
-        this.$message.error("删除失败" + err.message)
-      });
+     this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          CargosequenceService.deleteItem(obj.id, obj.version)
+          .then((res) => {
+            this.$message.success("删除成功");
+            this.schemeOrStore(this.changeActive, this.storeArmy, this.editProjectsInfo)
+          }).catch((err) => {
+            this.$message.error("删除失败" + err.message)
+          });
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          });          
+        });
    },
     handleSizeChangeOne(e) {
         this.mySelfPageSize = Number(e)
@@ -801,7 +816,8 @@ export default {
           grpId: this.storeArmy.id,
           storeCode: ele.code,
           storeName: ele.name,
-          storeOrder: idx + 1
+          storeId: ele.id,
+          storeOrder: idx + 1 + this.storeOption.length
         })
       })
       const postDataObj = {
@@ -810,8 +826,9 @@ export default {
       }
       CargosequenceService.addGrpItems(postDataObj)
       .then((res) => {
+        // console.log(this.changeActive, this.storeArmy, this.editProjectsInfo);
         this.$message.success("添加成功");
-        // this.schemeOrStore(this.changeActive, this.storeArmy, this.editProjectsInfo)
+        this.schemeOrStore(this.changeActive, this.storeArmy, this.editProjectsInfo)
       }).catch((err) => {
         this.$message.error("添加失败" + err.message)
       })
